# swagger.yaml
info: {}
paths:
    /config/:
        post:
            description: Add new config
            operationId: createConfig
            parameters:
                - description: 'name: body'
                  in: body
                  name: body
                  required: true
                  schema: {}
                  x-go-name: Body
            responses:
                "201":
                    $ref: '#/responses/ResponsePost'
                "400":
                    $ref: '#/responses/ErrorResponse'
                "415":
                    $ref: '#/responses/ErrorResponse'
            tags:
                - config
    /config/{id}/{version}/:
        delete:
            description: Delete config
            operationId: deleteConfig
            parameters:
                - description: Post ID
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: Id
            responses:
                "204":
                    $ref: '#/responses/NoContentResponse'
                "404":
                    $ref: '#/responses/ErrorResponse'
            tags:
                - config
        get:
            description: Get config by ID
            operationId: getConfigById
            parameters:
                - description: Post ID
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: Id
            responses:
                "200":
                    $ref: '#/responses/ResponsePost'
                "404":
                    $ref: '#/responses/ErrorResponse'
            tags:
                - config
    /configGroup/:
        post:
            description: Add new config group
            operationId: createConfigGroup
            parameters:
                - description: 'name: body'
                  in: body
                  name: body
                  required: true
                  schema: {}
                  x-go-name: Body
            responses:
                "201":
                    $ref: '#/responses/ResponsePost'
                "400":
                    $ref: '#/responses/ErrorResponse'
                "415":
                    $ref: '#/responses/ErrorResponse'
            tags:
                - configGroup
    /configGroup/{id}/:
        delete:
            description: Delete config group
            operationId: deleteConfigGroup
            parameters:
                - description: Post ID
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: Id
            responses:
                "204":
                    $ref: '#/responses/NoContentResponse'
                "404":
                    $ref: '#/responses/ErrorResponse'
            tags:
                - configGroup
        get:
            description: Get config group by ID
            operationId: getConfigGroupById
            parameters:
                - description: Post ID
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: Id
            responses:
                "200":
                    $ref: '#/responses/ResponsePost'
                "404":
                    $ref: '#/responses/ErrorResponse'
            tags:
                - configGroup
    /configGroups/:
        get:
            description: Get all config groups
            operationId: getConfigGroups
            responses:
                "200":
                    $ref: '#/responses/ResponsePost'
            tags:
                - configGroup
    /configs/:
        get:
            description: Get all configs
            operationId: getConfigs
            responses:
                "200":
                    $ref: '#/responses/ResponsePost'
            tags:
                - config
responses:
    ErrorResponse:
        description: ""
        headers:
            message:
                description: |-
                    Message of the error
                    in: string
                type: string
            status:
                description: |-
                    Error status code
                    in: int64
                format: int64
                type: integer
    NoContentResponse:
        description: ""
    ResponsePost:
        description: ""
        headers:
            id:
                description: |-
                    Id of the config
                    in: string
                type: string
            version:
                description: |-
                    Version of the config
                    in: string
                type: string
swagger: "2.0"